+++ scripts/rollback/rollback-2.12.0.sh
@@ -1,0 +1,151 @@
+#!/bin/bash
+# timestamp: 1627658431643
+
+npm dist-tag add @midwayjs/egg-layer@2.11.7 latest
+npm dist-tag add @midwayjs/egg-layer@2.12.0 beta
+tnpm dist-tag add @midwayjs/egg-layer@2.11.7 latest
+tnpm dist-tag add @midwayjs/egg-layer@2.12.0 latest
+npm dist-tag add @midwayjs/express-layer@2.11.6 latest
+npm dist-tag add @midwayjs/express-layer@2.12.0 beta
+tnpm dist-tag add @midwayjs/express-layer@2.11.6 latest
+tnpm dist-tag add @midwayjs/express-layer@2.12.0 latest
+npm dist-tag add @midwayjs/koa-layer@2.11.6 latest
+npm dist-tag add @midwayjs/koa-layer@2.12.0 beta
+tnpm dist-tag add @midwayjs/koa-layer@2.11.6 latest
+tnpm dist-tag add @midwayjs/koa-layer@2.12.0 latest
+npm dist-tag add @midwayjs/serverless-app@2.11.6 latest
+npm dist-tag add @midwayjs/serverless-app@2.12.0 beta
+tnpm dist-tag add @midwayjs/serverless-app@2.11.6 latest
+tnpm dist-tag add @midwayjs/serverless-app@2.12.0 latest
+npm dist-tag add @midwayjs/serverless-fc-starter@2.11.6 latest
+npm dist-tag add @midwayjs/serverless-fc-starter@2.12.0 beta
+tnpm dist-tag add @midwayjs/serverless-fc-starter@2.11.6 latest
+tnpm dist-tag add @midwayjs/serverless-fc-starter@2.12.0 latest
+npm dist-tag add @midwayjs/serverless-fc-trigger@2.11.6 latest
+npm dist-tag add @midwayjs/serverless-fc-trigger@2.12.0 beta
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@2.11.6 latest
+tnpm dist-tag add @midwayjs/serverless-fc-trigger@2.12.0 latest
+npm dist-tag add @midwayjs/serverless-scf-trigger@2.11.6 latest
+npm dist-tag add @midwayjs/serverless-scf-trigger@2.12.0 beta
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@2.11.6 latest
+tnpm dist-tag add @midwayjs/serverless-scf-trigger@2.12.0 latest
+npm dist-tag add @midwayjs/static-layer@2.11.6 latest
+npm dist-tag add @midwayjs/static-layer@2.12.0 beta
+tnpm dist-tag add @midwayjs/static-layer@2.11.6 latest
+tnpm dist-tag add @midwayjs/static-layer@2.12.0 latest
+npm dist-tag add @midwayjs/bootstrap@2.11.6 latest
+npm dist-tag add @midwayjs/bootstrap@2.12.0 beta
+tnpm dist-tag add @midwayjs/bootstrap@2.11.6 latest
+tnpm dist-tag add @midwayjs/bootstrap@2.12.0 latest
+npm dist-tag add @midwayjs/cache@2.11.6 latest
+npm dist-tag add @midwayjs/cache@2.12.0 beta
+tnpm dist-tag add @midwayjs/cache@2.11.6 latest
+tnpm dist-tag add @midwayjs/cache@2.12.0 latest
+npm dist-tag add @midwayjs/consul@2.11.6 latest
+npm dist-tag add @midwayjs/consul@2.12.0 beta
+tnpm dist-tag add @midwayjs/consul@2.11.6 latest
+tnpm dist-tag add @midwayjs/consul@2.12.0 latest
+npm dist-tag add @midwayjs/core@2.11.6 latest
+npm dist-tag add @midwayjs/core@2.12.0 beta
+tnpm dist-tag add @midwayjs/core@2.11.6 latest
+tnpm dist-tag add @midwayjs/core@2.12.0 latest
+npm dist-tag add @midwayjs/faas@2.11.6 latest
+npm dist-tag add @midwayjs/faas@2.12.0 beta
+tnpm dist-tag add @midwayjs/faas@2.11.6 latest
+tnpm dist-tag add @midwayjs/faas@2.12.0 latest
+npm dist-tag add @midwayjs/grpc@2.11.6 latest
+npm dist-tag add @midwayjs/grpc@2.12.0 beta
+tnpm dist-tag add @midwayjs/grpc@2.11.6 latest
+tnpm dist-tag add @midwayjs/grpc@2.12.0 latest
+npm dist-tag add midway-schedule@2.11.6 latest
+npm dist-tag add midway-schedule@2.12.0 beta
+tnpm dist-tag add midway-schedule@2.11.6 latest
+tnpm dist-tag add midway-schedule@2.12.0 latest
+npm dist-tag add midway@2.11.7 latest
+npm dist-tag add midway@2.12.0 beta
+tnpm dist-tag add midway@2.11.7 latest
+tnpm dist-tag add midway@2.12.0 latest
+npm dist-tag add @midwayjs/mock@2.11.6 latest
+npm dist-tag add @midwayjs/mock@2.12.0 beta
+tnpm dist-tag add @midwayjs/mock@2.11.6 latest
+tnpm dist-tag add @midwayjs/mock@2.12.0 latest
+npm dist-tag add @midwayjs/orm@2.11.6 latest
+npm dist-tag add @midwayjs/orm@2.12.0 beta
+tnpm dist-tag add @midwayjs/orm@2.11.6 latest
+tnpm dist-tag add @midwayjs/orm@2.12.0 latest
+npm dist-tag add @midwayjs/oss@2.11.7-beta.1 latest
+npm dist-tag add @midwayjs/oss@2.12.0 beta
+tnpm dist-tag add @midwayjs/oss@2.11.7-beta.1 latest
+tnpm dist-tag add @midwayjs/oss@2.12.0 latest
+npm dist-tag add @midwayjs/prometheus@2.11.6 latest
+npm dist-tag add @midwayjs/prometheus@2.12.0 beta
+tnpm dist-tag add @midwayjs/prometheus@2.11.6 latest
+tnpm dist-tag add @midwayjs/prometheus@2.12.0 latest
+npm dist-tag add @midwayjs/rabbitmq@2.11.6 latest
+npm dist-tag add @midwayjs/rabbitmq@2.12.0 beta
+tnpm dist-tag add @midwayjs/rabbitmq@2.11.6 latest
+tnpm dist-tag add @midwayjs/rabbitmq@2.12.0 latest
+npm dist-tag add @midwayjs/socketio@2.11.7 latest
+npm dist-tag add @midwayjs/socketio@2.12.0 beta
+tnpm dist-tag add @midwayjs/socketio@2.11.7 latest
+tnpm dist-tag add @midwayjs/socketio@2.12.0 latest
+npm dist-tag add @midwayjs/task@2.11.6 latest
+npm dist-tag add @midwayjs/task@2.12.0 beta
+tnpm dist-tag add @midwayjs/task@2.11.6 latest
+tnpm dist-tag add @midwayjs/task@2.12.0 latest
+npm dist-tag add @midwayjs/typegoose@2.11.6 latest
+npm dist-tag add @midwayjs/typegoose@2.12.0 beta
+tnpm dist-tag add @midwayjs/typegoose@2.11.6 latest
+tnpm dist-tag add @midwayjs/typegoose@2.12.0 latest
+npm dist-tag add @midwayjs/version@2.11.7 latest
+npm dist-tag add @midwayjs/version@2.12.0 beta
+tnpm dist-tag add @midwayjs/version@2.11.7 latest
+tnpm dist-tag add @midwayjs/version@2.12.0 latest
+npm dist-tag add @midwayjs/express@2.11.7 latest
+npm dist-tag add @midwayjs/express@2.12.0 beta
+tnpm dist-tag add @midwayjs/express@2.11.7 latest
+tnpm dist-tag add @midwayjs/express@2.12.0 latest
+npm dist-tag add @midwayjs/koa@2.11.6 latest
+npm dist-tag add @midwayjs/koa@2.12.0 beta
+tnpm dist-tag add @midwayjs/koa@2.11.6 latest
+tnpm dist-tag add @midwayjs/koa@2.12.0 latest
+npm dist-tag add @midwayjs/web@2.11.7 latest
+npm dist-tag add @midwayjs/web@2.12.0 beta
+tnpm dist-tag add @midwayjs/web@2.11.7 latest
+tnpm dist-tag add @midwayjs/web@2.12.0 latest
+npm dist-tag add @midwayjs/ws@2.11.6 latest
+npm dist-tag add @midwayjs/ws@2.12.0 beta
+tnpm dist-tag add @midwayjs/ws@2.11.6 latest
+tnpm dist-tag add @midwayjs/ws@2.12.0 latest
+
+# Changes:
+
+#  - @midwayjs/egg-layer: 2.11.7 => 2.12.0
+#  - @midwayjs/express-layer: 2.11.6 => 2.12.0
+#  - @midwayjs/koa-layer: 2.11.6 => 2.12.0
+#  - @midwayjs/serverless-app: 2.11.6 => 2.12.0
+#  - @midwayjs/serverless-fc-starter: 2.11.6 => 2.12.0
+#  - @midwayjs/serverless-fc-trigger: 2.11.6 => 2.12.0
+#  - @midwayjs/serverless-scf-trigger: 2.11.6 => 2.12.0
+#  - @midwayjs/static-layer: 2.11.6 => 2.12.0
+#  - @midwayjs/bootstrap: 2.11.6 => 2.12.0
+#  - @midwayjs/cache: 2.11.6 => 2.12.0
+#  - @midwayjs/consul: 2.11.6 => 2.12.0
+#  - @midwayjs/core: 2.11.6 => 2.12.0
+#  - @midwayjs/faas: 2.11.6 => 2.12.0
+#  - @midwayjs/grpc: 2.11.6 => 2.12.0
+#  - midway-schedule: 2.11.6 => 2.12.0
+#  - midway: 2.11.7 => 2.12.0
+#  - @midwayjs/mock: 2.11.6 => 2.12.0
+#  - @midwayjs/orm: 2.11.6 => 2.12.0
+#  - @midwayjs/oss: 2.11.7-beta.1 => 2.12.0
+#  - @midwayjs/prometheus: 2.11.6 => 2.12.0
+#  - @midwayjs/rabbitmq: 2.11.6 => 2.12.0
+#  - @midwayjs/socketio: 2.11.7 => 2.12.0
+#  - @midwayjs/task: 2.11.6 => 2.12.0
+#  - @midwayjs/typegoose: 2.11.6 => 2.12.0
+#  - @midwayjs/version: 2.11.7 => 2.12.0
+#  - @midwayjs/express: 2.11.7 => 2.12.0
+#  - @midwayjs/koa: 2.11.6 => 2.12.0
+#  - @midwayjs/web: 2.11.7 => 2.12.0
+#  - @midwayjs/ws: 2.11.6 => 2.12.0